refactor(reposicao): split DetalhesModal (707→168) em hook + componentes
Extrai a lógica do modal para o hook useDetalhesModal (2 queries, 5 mutations,
recalcularPedido, memos, flags de edição) e quebra o JSX em componentes
controlados: PortalStatusPanel, HistoricoAcoesPanel, CondicaoPagamentoPanel,
ItensTable e os 2 AlertDialogs (Remover/Descontinuar). Named export DetalhesModal
preservado (consumidor: AdminReposicaoPedidos). Comportamento 1:1; +19 testes.
